Separation of functionality [I] - captcha store for generating captchas outside of verify [I] - Button and Menu event files
