+++ src/events/interactionCreate.ts
@@ -1,16 +1,11 @@
-import { Interaction, MessageActionRow, MessageButton, MessageEmbed, MessageSelectMenu, TextChannel } from "discord.js";
+import type { Interaction, MessageActionRow } from "discord.js";
+import { handleJoinButtons } from "../buttons/joinEmbedButtons";
+import { handleRegenCaptcha } from "../buttons/regenCaptcha";
 import { getCommand } from "../commands";
-import {
-	captchaCache,
-	generateCaptchaForMember,
-	generateCaptchaImage,
-	generateFakeCaptchaList,
-} from "../commands/verify";
 import { log } from "../logger";
+import { handleCaptchaSelector } from "../menus/captchaSelector";
 import commandLogModel from "../models/commandLog.model";
-import { getState } from "../store/state";
 import type { Event } from "../types/event";
-import { embedCache } from "./guildMemberAdd";
 
 const event: Event = {
 	name: "interactionCreate",
@@ -38,58 +33,20 @@ const event: Event = {
 				const guild = interaction.guild;
 				const member = guild?.members.resolve(user.id);
 
+				if (!values || !guild || !member) {
+					interaction
+						.update({
+							components: [],
+							embeds: [],
+							content:
+								"Failed to handle interaction, please try again later. If the problem persists contact guild owner.",
+						})
+						.catch(err => log(err, "error"));
+					return;
+				}
+
 				if (name === "captchaSelector") {
-					const captcha = captchaCache.get(user.id) as string;
-					const selectedCaptcha = (values as string[])[0];
-
-					if (captcha === selectedCaptcha) {
-						interaction
-							.update({
-								components: [],
-								embeds: [],
-								content: "You have selected the correct captcha.",
-							})
-							.then(() => {
-								const messageId = embedCache.get(user.id);
-								const logChannel = guild?.channels.cache.get(
-									getState().config.joinLogChannelId as `${bigint}`
-								) as TextChannel;
-
-								if (logChannel && messageId) {
-									logChannel.messages
-										.fetch(messageId)
-										.then(message => {
-											message
-												.edit({
-													embeds: message.embeds,
-													components: [],
-												})
-												.then(() => {
-													embedCache.delete(user.id);
-												})
-												.catch(err => log(err, "error"));
-										})
-										.catch(err => log(err, "error"));
-								}
-							})
-							.catch(err => log(err, "error"));
-
-						if (member) {
-							member.roles.add(getState().config.memberRoleId).catch(console.warn.bind(console));
-						}
-					} else {
-						interaction
-							.update({
-								components: [],
-								embeds: [],
-								content: "You have selected an invalid captcha.",
-							})
-							.catch(err => log(err, "error"));
-
-						if (member) {
-							//member.kick("Invalid captcha.");
-						}
-					}
+					handleCaptchaSelector(interaction, values, guild, user, member);
 				}
 			} else if (interaction.isButton()) {
 				const name = interaction.customID;
@@ -101,109 +58,20 @@ const event: Event = {
 
 					if (member) {
 						if (name === "ban" || name === "kick") {
-							if (member.roles.cache.has(getState().config.staffRoleId)) {
-								if (components) {
-									const idToInteract = (components[0].components as MessageButton[]).filter(
-										button => button.customID === "memberId"
-									)[0].label as `${bigint}`;
-
-									if (idToInteract) {
-										const memberToInteract = guild.members.cache.get(idToInteract);
-
-										if (memberToInteract) {
-											if (name === "ban") {
-												memberToInteract
-													.ban({
-														reason: `Banned by ${member.displayName}[${member.id}] via join embed.`,
-													})
-													.then(() => {
-														interaction
-															.update({
-																embeds: interaction.message.embeds as MessageEmbed[],
-																components: [
-																	new MessageActionRow().addComponents(
-																		new MessageButton()
-																			.setCustomID("banned")
-																			.setStyle("SECONDARY")
-																			.setLabel(
-																				`Banned by ${member.displayName}[${member.id}]`
-																			)
-																			.setDisabled(true)
-																	),
-																],
-															})
-															.catch(err => log(err, "error"));
-													})
-													.catch(err => log(err, "error"));
-											} else if (name === "kick") {
-												memberToInteract
-													.kick(
-														`Kicked by ${member.displayName}[${member.id}] via join embed.`
-													)
-													.then(() => {
-														interaction
-															.update({
-																embeds: interaction.message.embeds as MessageEmbed[],
-																components: [
-																	new MessageActionRow().addComponents(
-																		new MessageButton()
-																			.setCustomID("kicked")
-																			.setStyle("SECONDARY")
-																			.setLabel(
-																				`Kicked by ${member.displayName}[${member.id}]`
-																			)
-																			.setDisabled(true)
-																	),
-																],
-															})
-															.catch(err => log(err, "error"));
-													})
-													.catch(err => log(err, "error"));
-											}
-										}
-									}
-								}
+							if (!components) {
+								interaction
+									.update({
+										components: [],
+										embeds: [],
+										content:
+											"Failed to handle interaction, please try again later. If the problem persists contact guild owner.",
+									})
+									.catch(err => log(err, "error"));
+								return;
 							}
+							handleJoinButtons(interaction, components as MessageActionRow[], guild, member, name);
 						} else if (name === "regenCaptcha") {
-							const captcha = generateCaptchaForMember(member.id);
-							const canvas = generateCaptchaImage(captcha);
-							const captchaData = generateFakeCaptchaList(captcha);
-
-							interaction.channel
-								?.send({
-									files: [
-										{
-											attachment: canvas.toBuffer(),
-											name: "captcha.png",
-										},
-									],
-								})
-								.then(async message => {
-									const captchaUrl = message.attachments.first()?.url;
-
-									await interaction.update({
-										embeds: [
-											new MessageEmbed().setTitle("Verification").setImage(captchaUrl ?? ""),
-										],
-										components: [
-											new MessageActionRow().addComponents(
-												new MessageButton()
-													.setCustomID("regenCaptcha")
-													.setLabel("Regenerate")
-													.setStyle("PRIMARY")
-											),
-											new MessageActionRow().addComponents(
-												new MessageSelectMenu()
-													.setCustomID("captchaSelector")
-													.setMinValues(1)
-													.setMaxValues(1)
-													.addOptions(captchaData)
-											),
-										],
-									});
-
-									message.delete().catch(err => log(err, "error"));
-								});
+							handleRegenCaptcha(interaction, member);
 						}
 					}
 				}
